pyproject.toml
[project]
name = "photoduplicates"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "imagehash>=4.3.2",
    "pillow>=12.0.0",
    "pyinstaller>=6.17.0",
    "pyside6>=6.10.1",
    "send2trash>=1.8.3",
]
